DO-NOT-UPSTREAM arm, firmware, LLVMLinux: replace naked function with hard coded parameters
As documented in GCC naked functions should only use Basic asm syntax. The
Extended asm or mixture of Basic asm and "C" code is not guarantee. Currently
this works because it was hard coded to follow and check GCC behavior for
arguments and register placement.

By replacing the naked function with an inline one it is possible to have the
same result in a more portable and reliable way. Register placement checks
are not needed since there is no hard code.

Not-signed-off-by: Behan Webster <behanw@converseincode.com>
Not-signed-off-by: Vinícius Tinti <viniciustinti@gmail.com>