MIPS: LLVMLinux: Fix an 'inline asm input/output type mismatch' error.
Change the type of csum_ipv6_magic's 'proto' argument from unsigned short to
__u32.

This fixes a type mismatch between the 'htonl(proto)' inline asm input, which
is __u32, and the 'proto' output, which is unsigned short.

This error gets reported by clang, but not by gcc.
The changed code can be compiled successfully by both gcc and clang.

Signed-off-by: Toma Tabacu <toma.tabacu@imgtec.com>